CR-2004-11-02-0: Enhancm.:  - added new commlib test binary: test_debug_client                             - added new parameter to statistic calculate function                               to force calculation for qping requests. Otherwise                               update statistics only once per minute                             - added debug client support into commlib                             - fixed bug when shutdown commlib with connected                               stream clients                             - fixed possible deadlock in commlib string list when                               the malloc() of an entry failed.                             - outcommented unused function:                               should_flush_event_client()                  Changed:   commlib, gdilib, qmaster                  Review:    pending
Issue number:
Obtained from:
Submitted by:
Reviewed by:

+++ source/libs/gdi/sge_any_request.h
@@ -69,7 +69,7 @@ enum {
 
 };
 
-const char* sge_dump_message_tag(int tag);
+const char* sge_dump_message_tag(unsigned long tag);
 int check_isalive(const char *masterhost);
 void prepare_enroll(const char *name);
 int sge_send_any_request(int synchron, u_long32 *mid, const char *rhost, const char *commproc, int id, sge_pack_buffer *pb, int tag, u_long32 response_id, lList **alpp);
